refactor useSettings into own file
