+++ src/index.js
@@ -1,1 +1,1 @@
-import './settings';
+import './settings/SettingsPage'
